with imageable

+++ db/migrate/20180131095818_remove_mobilenumber_from_member.rb
@@ -1,5 +1,5 @@
 class RemoveMobilenumberFromMember < ActiveRecord::Migration[5.1]
   def change
-    remove_column :members, :mobileNumber, :float
+    remove_column :members, :mobileNumber, :bigint
   end
 end
